Change lifespan constructor parameter type from int to float to match its declared type.

+++ panda/src/particlesystem/pointParticle.cxx
@@ -24,7 +24,7 @@
 // Description : simple constructor
 ////////////////////////////////////////////////////////////////////
 PointParticle::
-PointParticle(int lifespan, bool alive) :
+PointParticle(float lifespan, bool alive) :
   BaseParticle(lifespan, alive) {
   set_oriented(false);
 }
